requirements.txt
git+https://github.com/nghorbani/human_body_prior.git@SOMA
git+https://github.com/nghorbani/body_visualizer
git+https://github.com/nghorbani/configer.git
imageio
mkl
mkl-service
mkl_fft
mkl_random
numpy
pyOpenSSL
pillow
scikit-image
scipy
setuptools
loguru
omegaconf
six
tk
toolz
wheel
pytorch3d
notifiers
tqdm
opencv-python
c3d
chumpy
ipython
markdown
omegaconf
pandas
pycodestyle
pytorch-lightning
scikit-learn
tensorboard
threadpoolctl
setuptools
trimesh
colour
seaborn
tables
transforms3d
xlsxwriter
jupyterlab